Extracted class Ui
