+++ src/main/java/Main.java
@@ -1,45 +1,19 @@
 import doraemonexceptions.EmptyListException;
 import doraemonexceptions.InValidCommandException;
 import doraemonexceptions.IsEmptyException;
+import todo.Todo;
+import ui.Ui;
+
 import java.util.Scanner;
 import java.util.ArrayList;
 import java.io.File;
-//import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 
 
 public class Main {
     public static final String PATHNAME = "src/data/prevData";
-    public static void printLine() {
-        System.out.println("____________________________________________________________");
-    }
 
-    public static void printGreetings() {
-        printLine();
-        System.out.println("Hello! I'm Doraemon!");
-        System.out.println("What can I do for you?");
-        printLine();
-    }
-
-    public static void printMessage(Todo task, int taskNum) {
-        System.out.println("Got it. I've added this task:");
-        System.out.println("  " + task.formatTask());
-        System.out.println("Now you have " + taskNum + " tasks in the list.");
-        printLine();
-    }
-
-    public static void printDelMessage(Todo task, int taskNum) {
-        System.out.println("Got it. I've deleted this task:");
-        System.out.println("  " + task.formatTask());
-        System.out.println("Now you have " + taskNum + " tasks in the list.");
-        printLine();
-    }
-
-    public static void printMarked(Todo task) {
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println(task.formatTask());
-    }
     private static void addTodo(ArrayList<Todo> list, String task, int taskNum) {
         list.add(taskNum, new Todo(task));
     }
@@ -60,13 +34,13 @@ public class Main {
         list.add(taskNum, new Event(description, start, end)) ;
     }
 
-    public static int loadData(ArrayList<Todo> list) throws IOException {
+    public static int loadData(ArrayList<Todo> list, Ui ui) throws IOException {
         File storedFile = new File(PATHNAME);
         int taskNum = 0;
         if (!storedFile.exists()) {
             new File(storedFile.getParent()).mkdirs();
             System.out.println("No previous data was found, a new file is created.");
-            printLine();
+            ui.printLine();
         } else {
             Scanner dataInput = new Scanner(storedFile);
             while (dataInput.hasNext()) {
@@ -100,15 +74,16 @@ public class Main {
     }
 
     public static void main(String[] args) {
+        Ui ui = new Ui();
         ArrayList<Todo> list = new ArrayList<>();
         int taskNum = 0;
         Scanner in = new Scanner(System.in);
-        printGreetings();
+        ui.printGreetings();
         try {
-            taskNum = loadData(list);
+            taskNum = loadData(list, ui);
         } catch (IOException e) {
             System.out.println("Something went wrong, please try again!");
-            printLine();
+            ui.printLine();
         }
         String temp = in.nextLine();
         while (!temp.equals("bye")) {
@@ -120,36 +95,36 @@ public class Main {
                     for (int i = 0; i < taskNum; i++) {
                         System.out.println(i + 1 + "." + list.get(i).formatTask());
                     }
-                    printLine();
+                    ui.printLine();
                 } else if (temp.startsWith("todo")) {
                     temp = temp.substring(4); //trim off command
                     addTodo(list, temp, taskNum);
-                    printMessage(list.get(taskNum), taskNum + 1);
+                    ui.printMessage(list.get(taskNum), taskNum + 1);
                     writeTask(list.get(taskNum));
                     taskNum++;
                 } else if (temp.startsWith("deadline")) {
                     temp = temp.substring(8); //trim off command
                     addDeadline(list, temp, taskNum);
-                    printMessage(list.get(taskNum), taskNum + 1);
+                    ui.printMessage(list.get(taskNum), taskNum + 1);
                     writeTask(list.get(taskNum));
                     taskNum++;
                 } else if (temp.startsWith("event")) {
                     temp = temp.substring(5); //trim off command
                     addEvent(list, temp, taskNum);
-                    printMessage(list.get(taskNum), taskNum + 1);
+                    ui.printMessage(list.get(taskNum), taskNum + 1);
                     writeTask(list.get(taskNum));
                     taskNum++;
                 } else if (temp.startsWith("mark")) {
                     temp = temp.substring(5); //trim "mark"
                     int indexToMark = Integer.parseInt(temp) - 1;
                     list.get(indexToMark).markStatus();
-                    printMarked(list.get(indexToMark));
-                    printLine();
+                    ui.printMarked(list.get(indexToMark));
+                    ui.printLine();
                 } else if (temp.startsWith("delete")){
                     temp = temp.substring(7); //trim delete
                     int indexToDel = Integer.parseInt(temp) - 1;
                     taskNum--;
-                    printDelMessage(list.get(indexToDel), taskNum);
+                    ui.printDelMessage(list.get(indexToDel), taskNum);
                     list.remove(indexToDel);
                 } else if (temp.isEmpty()) {
                     throw new IsEmptyException();
@@ -159,21 +134,20 @@ public class Main {
                 temp = in.nextLine();
             } catch (IsEmptyException e) {
                 System.out.println("Please type something! Try: todo/deadline/event to add a record!");
-                printLine();
+                ui.printLine();
                 temp = in.nextLine();
             } catch (InValidCommandException e) {
                 System.out.println("Invalid Command! Try: todo/deadline/event to add a record!");
-                printLine();
+                ui.printLine();
                 temp = in.nextLine();
             } catch (EmptyListException e) {
                 System.out.println("The list is empty! Please add something!");
-                printLine();
+                ui.printLine();
                 temp = in.nextLine();
             } catch (IOException e) {
                 System.out.println("Something went wrong: " + e.getMessage());
             }
         }
-        System.out.println("Bye. Have a great day!");
-        printLine();
+        ui.printExit();
     }
 }
